Move pending catches to Caught screen
